storage: add minimum record version hint
