+++ pkg/storage/querier.go
@@ -3,20 +3,15 @@ package storage
 import (
 	"context"
 	"encoding/json"
-	"strconv"
 
-	"github.com/google/uuid"
 	grpc "google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	status "google.golang.org/grpc/status"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
-	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 
-	"github.com/pomerium/pomerium/pkg/cryptutil"
 	"github.com/pomerium/pomerium/pkg/grpc/databroker"
 	"github.com/pomerium/pomerium/pkg/grpcutil"
-	"github.com/pomerium/pomerium/pkg/protoutil"
 )
 
 // A Querier is a read-only subset of the client methods
@@ -50,135 +45,9 @@ func WithQuerier(ctx context.Context, querier Querier) context.Context {
 	return context.WithValue(ctx, querierKey{}, querier)
 }
 
-type staticQuerier struct {
-	records map[string]RecordCollection
-}
-
-// NewStaticQuerier creates a Querier that returns statically defined protobuf records.
-func NewStaticQuerier(msgs ...proto.Message) Querier {
-	getter := &staticQuerier{records: make(map[string]RecordCollection)}
-	for _, msg := range msgs {
-		record, ok := msg.(*databroker.Record)
-		if !ok {
-			record = NewStaticRecord(protoutil.NewAny(msg).TypeUrl, msg)
-		}
-		c, ok := getter.records[record.Type]
-		if !ok {
-			c = NewRecordCollection()
-			getter.records[record.Type] = c
-		}
-		c.Put(record)
-	}
-	return getter
-}
-
-// NewStaticRecord creates a new databroker Record from a protobuf message.
-func NewStaticRecord(typeURL string, msg proto.Message) *databroker.Record {
-	data := protoutil.NewAny(msg)
-	record := new(databroker.Record)
-	record.ModifiedAt = timestamppb.Now()
-	record.Version = cryptutil.NewRandomUInt64()
-	record.Id = uuid.New().String()
-	record.Data = data
-	record.Type = typeURL
-	if hasID, ok := msg.(interface{ GetId() string }); ok {
-		record.Id = hasID.GetId()
-	}
-	if hasVersion, ok := msg.(interface{ GetVersion() string }); ok {
-		if v, err := strconv.ParseUint(hasVersion.GetVersion(), 10, 64); err == nil {
-			record.Version = v
-		}
-	}
-
-	var jsonData struct {
-		ID      string `json:"id"`
-		Version string `json:"version"`
-	}
-	bs, _ := protojson.Marshal(msg)
-	_ = json.Unmarshal(bs, &jsonData)
-
-	if jsonData.ID != "" {
-		record.Id = jsonData.ID
-	}
-	if jsonData.Version != "" {
-		if v, err := strconv.ParseUint(jsonData.Version, 10, 64); err == nil {
-			record.Version = v
-		}
-	}
-
-	return record
-}
-
-func (q *staticQuerier) InvalidateCache(_ context.Context, _ *databroker.QueryRequest) {}
-
 // Query queries for records.
 func (q *staticQuerier) Query(_ context.Context, req *databroker.QueryRequest, _ ...grpc.CallOption) (*databroker.QueryResponse, error) {
 	return QueryRecordCollections(q.records, req)
-}
-
-type clientQuerier struct {
-	client databroker.DataBrokerServiceClient
-}
-
-// NewQuerier creates a new Querier that implements the Querier interface by making calls to the databroker over gRPC.
-func NewQuerier(client databroker.DataBrokerServiceClient) Querier {
-	return &clientQuerier{client: client}
-}
-
-func (q *clientQuerier) InvalidateCache(_ context.Context, _ *databroker.QueryRequest) {}
-
-// Query queries for records.
-func (q *clientQuerier) Query(ctx context.Context, in *databroker.QueryRequest, opts ...grpc.CallOption) (*databroker.QueryResponse, error) {
-	return q.client.Query(ctx, in, opts...)
-}
-
-type cachingQuerier struct {
-	q     Querier
-	cache Cache
-}
-
-// NewCachingQuerier creates a new querier that caches results in a Cache.
-func NewCachingQuerier(q Querier, cache Cache) Querier {
-	return &cachingQuerier{
-		q:     q,
-		cache: cache,
-	}
-}
-
-func (q *cachingQuerier) InvalidateCache(ctx context.Context, in *databroker.QueryRequest) {
-	key, err := (&proto.MarshalOptions{
-		Deterministic: true,
-	}).Marshal(in)
-	if err != nil {
-		return
-	}
-	q.cache.Invalidate(key)
-	q.q.InvalidateCache(ctx, in)
-}
-
-func (q *cachingQuerier) Query(ctx context.Context, in *databroker.QueryRequest, opts ...grpc.CallOption) (*databroker.QueryResponse, error) {
-	key, err := MarshalQueryRequest(in)
-	if err != nil {
-		return nil, err
-	}
-
-	rawResult, err := q.cache.GetOrUpdate(ctx, key, func(ctx context.Context) ([]byte, error) {
-		res, err := q.q.Query(ctx, in, opts...)
-		if err != nil {
-			return nil, err
-		}
-		return MarshalQueryResponse(res)
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	var res databroker.QueryResponse
-	err = proto.Unmarshal(rawResult, &res)
-	if err != nil {
-		return nil, err
-	}
-	return &res, nil
 }
 
 // MarshalQueryRequest marshales the query request.
@@ -208,6 +77,9 @@ func GetDataBrokerRecord(
 		Type:  recordType,
 		Limit: 1,
 	}
+	if lowestRecordVersion > 0 {
+		req.MinimumRecordVersionHint = proto.Uint64(lowestRecordVersion)
+	}
 	req.SetFilterByIDOrIndex(recordID)
 
 	res, err := q.Query(ctx, req, grpc.WaitForReady(true))
@@ -217,23 +89,6 @@ func GetDataBrokerRecord(
 	if len(res.GetRecords()) == 0 {
 		return nil, ErrNotFound
 	}
-
-	// if the current record version is less than the lowest we'll accept, invalidate the cache
-	if res.GetRecords()[0].GetVersion() < lowestRecordVersion {
-		q.InvalidateCache(ctx, req)
-	} else {
-		return res.GetRecords()[0], nil
-	}
-
-	// retry with an up to date cache
-	res, err = q.Query(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	if len(res.GetRecords()) == 0 {
-		return nil, ErrNotFound
-	}
-
 	return res.GetRecords()[0], nil
 }
